Refactor(store): modularize editor store actions and reduce main file size
