+++ frontend_react/src/components/CategoryEditorV2/hooks/store/index.ts
@@ -1,0 +1,13 @@
+// hooks/store/index.ts
+// ------------------------------------------------------
+// 將所有 editorStore actions 統一匯出，供 useEditorStore.ts 使用
+// ------------------------------------------------------
+
+export { getConfigActions } from './editorStoreConfigActions';
+export { getConfigRemovalActions } from './editorStoreConfigRemovalActions';
+export { getBadgeActions } from './editorStoreBadgeActions';
+export { getFilterActions } from './editorStoreFilterActions';
+export { getCustomKeywordActions } from './editorStoreCustomKeywordActions';
+export { getSelectionActions } from './editorStoreSelectionActions';
+export { getCategoryActions } from './editorStoreCategoryActions';
+export { getSyncActions } from './editorStoreSyncActions';
