+++ frontend_react/src/components/CategoryEditorV2/types/editor.ts
@@ -51,6 +51,8 @@ export interface Video {
   badges?: Badge[];
 }
 
+export type KeywordSource = 'bracket' | 'frequency' | 'game' | 'custom';
+
 export interface EditorState {
   selectedBySource: {
     bracket: Set<string>;
